fix: retain the copied selection on remote panes, and record two SDK walls
The remote clipboard arm cleared the selection on a successful copy while the
local arm kept it. Local wins: keeping the highlight is the convention in
every Mac terminal these users arrive from, and it is the only confirmation
that the copy took the range they meant — clearing it deletes the evidence at
the moment you would check it. Typing, a new selection, and restart still
clear it on both sides.

Also widens the shortcut latch u32 -> u64 across all three of its views.
Every bit was spoken for; the next chord added would have been silently
dropped.

The rest of this commit is findings, not features, because two things that
were asked for turned out to be blocked below this repo — and a chord that
half-works is worse than one that does not exist.

MULTIPLE WINDOWS are blocked in the SDK, verified in the running app rather
than argued. The window ceiling is real and fine (5: one scene window plus
four model-declared secondaries, refused cleanly at six). The blocker is
painting: every terminal cell comes from `chrome.build`, the SDK installs an
app-owned chrome display list for the MAIN canvas only, and `ChromeOptions
.build` takes no window label so it could not paint another window even if
it were called. A spike produced a second window that rendered the tab strip,
the divider and correct widget bounds with ZERO terminal cells — two
screenshots at the same instant, one live and one empty. `on_frame` and the
webview pump are main-canvas-only for the same reason, so those terminals
would never be resized either. The spike was backed out and `cmd+N` was not
shipped.

FULLSCREEN is blocked the same way. `PlatformServices` has focus/close/
minimize/show and REPORTS `WindowInfo.fullscreen`, but has no verb that
enters or leaves it. The stock View menu carries one, but supplying any
custom menu replaces the whole bar and every custom item routes through one
command action. The half that usually breaks does not: a driven resize
reflowed both panes and reached zsh, so only the trigger is missing. The
green traffic-light button still works.

EASING THE CHROME BAND is refused on two independent grounds. The SDK's
layout tween animates the widget tree only, while the painter, hit targets
and PTY pump all derive from `workspaceChrome` — so an eased tree over a
stepped content rect leaves all three disagreeing for the length of the
animation. Easing the extent instead keeps them in lockstep and moves
`content.height` every frame, which costs three SIGWINCHes per pane per
reveal over a 50pt band and a 17.5pt cell. The band's CONTENTS can be
animated freely; only its extent is load-bearing.

302 pass / 5 skip.

+++ src/cockpit/model.zig
@@ -215,7 +215,7 @@ pub const Model = struct {
     browser_page: BrowserPage = .github,
     browser_navigation_token: u64 = 0,
     focused: bool = true,
-    consumed_shortcut_keys_held: u32 = 0,
+    consumed_shortcut_keys_held: u64 = 0,
     held_terminal_keys: [max_held_terminal_keys]HeldTerminalKey = [_]HeldTerminalKey{.{}} ** max_held_terminal_keys,
     pointer_captures: [max_pointer_captures]PointerCapture = [_]PointerCapture{.{ .terminal_id = .terminal_1 }} ** max_pointer_captures,
     copy_inflight: bool = false,
